pyproject.toml
[project]
name = "chatbot-colombina"
version = "0.1.0"
description = "Chatbot for Colombina's document repository"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "chromadb",
    "langchain",
    "langchain-core",
    "langchain-openai",
    "langchain-ollama",
    "langchain-chroma",
    "langchain-text-splitters",
    "langgraph",
    "streamlit",
    "beautifulsoup4",
    "lxml",
    "openpyxl",
    "pandas",
    "pdfplumber",
    "pypdf2",
    "requests",
    "selenium",
    "urllib3",
    "webdriver-manager",
    "fastapi",
    "uvicorn",
    "pydantic"
]